create a router file and render it inside app
