+++ package.json
@@ -1,8 +1,8 @@
 {
-  "homepage": "https://prakritip0.github.io/task-4 ",
+  "homepage": "/",
   "name": "task-4",
   "version": "0.1.0",
-  "private": false,
+  "private": true,
   "type": "module",
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
